refactor structure and move traits to core
