+++ core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "anunaya-rollup-core"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+
+serde = {workspace = true, features = ["derive"] }
+alloy = { workspace = true }
+serde_json = { workspace = true }
