+++ core/src/traits/block.rs
@@ -1,10 +1,17 @@
 #![allow(dead_code)]
 use std::fmt::Debug;
+use super::*;
 
-pub trait HasherT: Sync + Send {
-    type Output: AsRef<[u8]> + Send + Sync + Clone + Debug + Copy;
-    fn hash(s: &[u8]) -> Self::Output;
+pub trait BlockT: Clone + Send + Sync + Debug + 'static {
+    type Transaction: SignedTransactionT;
+    type BlockHeader: BlockHeaderT;
+
+    fn header(&self) -> &Self::BlockHeader;
+    fn transactions(&self) -> &[Self::Transaction];
+    fn new(header: Self::BlockHeader, transactions: Vec<Self::Transaction>) -> Self;
 }
+
+
 
 pub trait BlockHeaderT: Clone + Send + Sync + Debug + 'static {
     // Header number
@@ -35,13 +42,3 @@ pub trait BlockHeaderT: Clone + Send + Sync + Debug + 'static {
     }
 }
 
-pub trait SignedTransactionT: Clone + Send + Sync + Debug + 'static {}
-
-pub trait BlockT: Clone + Send + Sync + Debug + 'static {
-    type Transaction: SignedTransactionT;
-    type BlockHeader: BlockHeaderT;
-
-    fn header(&self) -> &Self::BlockHeader;
-    fn transactions(&self) -> &[Self::Transaction];
-    fn new(header: Self::BlockHeader, transactions: Vec<Self::Transaction>) -> Self;
-}
