+++ core/src/traits/mod.rs
@@ -1,0 +1,9 @@
+mod block;
+mod hasher;
+mod signed_tx;
+mod state_machine;
+
+// Re-export traits for easier access
+pub use block::*;
+pub use hasher::*;
+pub use signed_tx::*;
