+++ core/src/traits/signed_tx.rs
@@ -1,0 +1,2 @@
+use std::fmt::Debug;
+pub trait SignedTransactionT: Clone + Send + Sync + Debug + 'static {}
