+++ core/src/traits/state_machine.rs
@@ -1,0 +1,8 @@
+
+pub trait DataAvailablityLayer {
+    
+}
+
+pub trait StateTransitionFunction<D: DataAvailablityLayer> {
+    
+}
